Commit 047aec0

+++ drivers/power/supply/qcom/smb1398-charger.c
@@ -2732,7 +2732,7 @@ static int smb1398_probe(struct platform_device *pdev)
 	chip->disabled = true;
 	platform_set_drvdata(pdev, chip);
 
-	chip->div2_cp_role = (int)of_device_get_match_data(chip->dev);
+	chip->div2_cp_role = (int)(unsigned long)of_device_get_match_data(chip->dev);
 	switch (chip->div2_cp_role) {
 	case DIV2_CP_MASTER:
 	case SMB1394_DIV2_CP_PRY:
